Split EndAlignment cop into EndAlignment and BlockAlignment cops.
