+++ lib/rubocop.rb
@@ -14,6 +14,7 @@ require 'rubocop/cop/cop'
 require 'rubocop/cop/commissioner'
 
 require 'rubocop/cop/lint/assignment_in_condition'
+require 'rubocop/cop/lint/block_alignment'
 require 'rubocop/cop/lint/empty_ensure'
 require 'rubocop/cop/lint/end_alignment'
 require 'rubocop/cop/lint/end_in_method'
