+++ lib/rubocop/cop/lint/block_alignment.rb
@@ -1,0 +1,151 @@
+# encoding: utf-8
+
+module Rubocop
+  module Cop
+    module Lint
+      # This cop checks whether the end keywords are aligned properly for do
+      # end blocks.
+      #
+      # @example
+      #
+      #   variable = lambda do |i|
+      #     i
+      #   end
+      class BlockAlignment < Cop
+        MSG = 'end at %d, %d is not aligned with %s at %d, %d'
+
+        def initialize
+          super
+          @inspected_blocks = []
+        end
+
+        def on_block(node)
+          return if already_processed_node?(node)
+          method, = *node
+          start_node = method.loc.expression.source =~ /\n/ ? method : node
+          check_block_alignment(start_node.loc.expression, node.loc)
+        end
+
+        def on_and(node)
+          return if already_processed_node?(node)
+
+          _left, right = *node
+          if right.type == :block
+            check_block_alignment(node.loc.expression, right.loc)
+            @inspected_blocks << right
+          end
+        end
+
+        alias_method :on_or, :on_and
+
+        def on_lvasgn(node)
+          _, children = *node
+          process_block_assignment(node, children)
+        end
+
+        alias_method :on_ivasgn,   :on_lvasgn
+        alias_method :on_cvasgn,   :on_lvasgn
+        alias_method :on_gvasgn,   :on_lvasgn
+        alias_method :on_and_asgn, :on_lvasgn
+        alias_method :on_or_asgn,  :on_lvasgn
+
+        def on_casgn(node)
+          _, _, children = *node
+          process_block_assignment(node, children)
+        end
+
+        def on_op_asgn(node)
+          variable, _op, args = *node
+          process_block_assignment(variable, args)
+        end
+
+        def on_send(node)
+          _receiver, _method, *args = *node
+          process_block_assignment(node, args.last)
+        end
+
+        def on_masgn(node)
+          variables, args = *node
+          process_block_assignment(variables, args)
+        end
+
+        private
+
+        def process_block_assignment(begin_node, other_node)
+          return unless other_node
+
+          block_node = find_block_node(other_node)
+          return unless block_node.type == :block
+
+          # If the block is an argument in a function call, align end with
+          # the block itself, and not with the function.
+          if begin_node.type == :send
+            _receiver, method, *_args = *begin_node
+            begin_node = block_node if method.to_s =~ /^\w+$/
+          end
+
+          # Align with the expression that is on the same line
+          # where the block is defined
+          if begin_node.type != :mlhs && block_is_on_next_line?(begin_node,
+                                                                block_node)
+            return
+          end
+          return if already_processed_node?(block_node)
+
+          @inspected_blocks << block_node
+          check_block_alignment(begin_node.loc.expression, block_node.loc)
+        end
+
+        def find_block_node(node)
+          while [:send, :lvasgn].include?(node.type)
+            n = case node.type
+                when :send
+                  find_block_or_send_node(node) || break
+                when :lvasgn
+                  _variable, value = *node
+                  value
+                end
+            node = n if n
+          end
+          node
+        end
+
+        def find_block_or_send_node(send_node)
+          receiver, _method, args = *send_node
+          [receiver, args].find do |subnode|
+            subnode && [:block, :send].include?(subnode.type)
+          end
+        end
+
+        def check_block_alignment(start_loc, block_loc)
+          match = start_loc.source.match(/\n(\s*)((end)?\.\S+)\Z/)
+          if match
+            start_line = start_loc.line + start_loc.source.count("\n")
+            start_column = match.captures[0].length
+            start_source = match.captures[1]
+          else
+            start_line = start_loc.line
+            start_column = start_loc.column
+            start_source = start_loc.source.lines.to_a.first.chomp
+          end
+          end_loc = block_loc.end
+          if block_loc.begin.line != end_loc.line &&
+              start_column != end_loc.column
+            add_offence(:warning,
+                        end_loc,
+                        sprintf(MSG, end_loc.line, end_loc.column,
+                                start_source, start_line, start_column))
+          end
+        end
+
+        def already_processed_node?(node)
+          @inspected_blocks.include?(node)
+        end
+
+        def block_is_on_next_line?(begin_node, block_node)
+          begin_node.loc.line != block_node.loc.line
+        end
+      end
+    end
+  end
+end
